Components refactorization
