+++ source/assets/components/recipeCard/iconRecipeCard.css
@@ -1,0 +1,95 @@
+/** iconRecipeCard.css **/
+.circle-1dollar {
+  background: #27ba4e;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  text-align: center;
+  line-height: 30px;
+  vertical-align: middle;
+  padding: 10px;
+  margin-left: 10px;
+}
+
+#circle-2dollar {
+  background: #27ba4e;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  text-align: center;
+  line-height: 30px;
+  vertical-align: middle;
+  padding: 3px;
+  margin: 10px;
+}
+
+#circle-3dollar {
+  background: #27ba4e;
+  font-size: small;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  text-align: center;
+  line-height: 30px;
+  vertical-align: middle;
+  padding: 3px;
+  margin: 10px;
+}
+
+.smiley-face,
+.neutral-face,
+.sad-face {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  text-align: center;
+  line-height: 30px;
+  vertical-align: middle;
+  padding: 10px;
+  margin-left: 10px;
+}
+
+.smiley-face,
+.less-than-15-min {
+  background-color: lightgreen;
+}
+
+.neutral-face,
+.between-15-and-45-min {
+  background-color: yellow;
+}
+
+.sad-face,
+.over-45-min {
+  background-color: lightsalmon;
+}
+
+.one-serving,
+.two-servings,
+.three-plus-servings,
+.less-than-15-min,
+.between-15-and-45-min,
+.over-45-min {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  text-align: center;
+  vertical-align: middle;
+  padding: 10px;
+  margin-left: 10px;
+}
+
+.one-serving {
+  line-height: 30px;
+}
+
+.two-servings,
+.three-plus-servings {
+  line-height: 20px;
+}
+
+.one-serving,
+.two-servings,
+.three-plus-servings {
+  background-color: lightblue;
+}
