+++ source/index.html
@@ -4,8 +4,8 @@
     <title>Recipe App</title>
 
     <!-- Scripts -->
-    <script src="assets/components/recipeCard.js" type="module"></script>
-    <script src="assets/components/recipeViewer.js" type="module"></script>
+    <script src="assets/components/recipeCard/recipeCard.js" type="module"></script>
+    <script src="assets/components/recipeViewer/recipeViewer.js" type="module"></script>
     <script src="assets/scripts/main.js" type="module"></script>
     <script src="assets/scripts/spoonacular.js" type="module"></script>
     <script src="assets/scripts/filter.js" type="module"></script>
@@ -15,6 +15,7 @@
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+      type="text/css"
     />
   </head>
 
@@ -193,12 +194,6 @@
         <recipe-viewer></recipe-viewer>
       </section>
     </main>
-    <!-- Bootstrap -->
-    <script
-      src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-      integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-      crossorigin="anonymous"
-    ></script>
     <!-- Font Awesome -->
     <script src="https://kit.fontawesome.com/0e68289b8e.js" crossorigin="anonymous"></script>
   </body>
